feat: add document management and harden hours accounting flow

+++ src/main/java/com/LigaAcademic/AcademicProject/DTO/RegistroAtividadesResponseDTO.java
@@ -1,12 +1,13 @@
 package com.ligaacademic.academicproject.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 public record RegistroAtividadesResponseDTO(
 
         Long id,
-        float horas,
+        BigDecimal horas,
         List<String> participantes,
         String setorAtividade,
         String descAtividade,
